Codechange: Don't use enums for non-enumerated values. (#13031)
In the past we have used enums to hold an arbitrary values. These values
are not enumerated types, so make them constants instead.

+++ src/spritecache.cpp
@@ -661,7 +661,7 @@ bool LoadNextSprite(int load_index, SpriteFile &file, uint file_sprite_id)
 
 	if (type == SpriteType::Invalid) return false;
 
-	if (load_index >= MAX_SPRITES) {
+	if (static_cast<uint>(load_index) >= MAX_SPRITES) {
 		UserError("Tried to load too many sprites (#{}; max {})", load_index, MAX_SPRITES);
 	}
